text overflow and wrapping working. also added worst animation system to support test for the overflow and wrapping. still a few bugs but moving on

+++ src/gde/engine/platform.go
@@ -1,24 +1,24 @@
 package engine
 
 type Platform struct {
-	ScreenW     int32
-	ScreenH     int32
-	RenderW     int32
-	RenderH     int32
+	ScreenW     float32
+	ScreenH     float32
+	RenderW     float32
+	RenderH     float32
 	AspectRatio float32
 	Landscape   bool
 }
 
-func (p *Platform) NewPlatform(sw int32, sh int32, rw int32, rh int32) {
+func (p *Platform) NewPlatform(sw float32, sh float32, rw float32, rh float32) {
 	p.ScreenW = sw
 	p.ScreenH = sh
 	p.RenderW = rw
 	p.RenderH = rh
 	if p.RenderW > p.RenderH {
 		p.Landscape = true
-		p.AspectRatio = float32(p.RenderW) / float32(p.RenderH)
+		p.AspectRatio = p.RenderW / p.RenderH
 	} else {
 		p.Landscape = false
-		p.AspectRatio = float32(p.RenderH) / float32(p.RenderW)
+		p.AspectRatio = p.RenderH / p.RenderW
 	}
 }
